tmc: Move common TMC stepper driver code to new file tmc.py
Move common code from tmc2130.py to tmc.py.
